deps: replace oras-go with go-containerregistry
oras-go v1 has often caused issues due to it depending on old versions
of packages, and blocking dependency updates.

The current v2 of oras-go has a completely different structure, and
would require a comprehensive rewrite of our oras related code.

As we are using go-containerregistry for handling oci-sif push/pull,
we can use it for oras sif push/pull and drop oras-go. This also
requires a rewrite, but should ease the dependency situation going
forward.

Note that we now have to explicitly handle adding and removing items
from .singularity/docker-config.json in registry login/logout, as we
were using some oras-go v1 functions there before. go-cr will perform
unauthenticated GET against registries with a public `/v2/` endpoint,
so for the moment passwords are only verified on login when the
registry is fully private. Noted in changelog, may be addressed in
future.

Fixes #1262